Extract to screenblock file
